\name{landim1}
\alias{landim1}

\docType{data}
\title{
geoR's landim1 dataset
}
\description{
Dataset originally from geoR package. Dataframe with 38 locations with information related to easting (\code{EW}), northing (\code{NS}), and two generic numeric variables (\code{A} and \code{B}).
}
\usage{data("landim1")}
\format{
The format is:
 chr "landim1"
}
\details{
Run \code{str(landim1)}, \code{data(landim1)} or \code{summary(landim1)} to see more details about the dataset.
}
\source{
geoR package \url{https://CRAN.r-project.org/package=geoR}.
}

