\name{B3. Summary}
\alias{B3. Summary}
\alias{summary}
\alias{summary,sssFit-method}
\title{
Summary of the estimated model from \code{scp} object.
}
\description{
Report estimated coefficients, standard errors and t-tests for parametric effects of the semiparametric model from a \code{scp} object.
}
\usage{
\S4method{summary}{sssFit}(object,alpha=0.05)
}
\arguments{
	\item{object}{
		sssFit object from \link{scp}.
	}
	\item{alpha}{
		numeric \eqn{[0,1]}. The level for confidence intervals.
	}

}
\author{
Mario A. Martinez Araya, \email{r@marioma.me}
}
%% ~Make other sections like Warning with \section{Warning }{....} ~ 

